03_RNN_raincast/requirements.txt
torch
torchvision
pytorch-lightning
h5py
matplotlib
wandb
hydra-core
omegaconf
numpy
pandas
